+++ pizzeria-console-objet-java8-maven-multimodule/pizzeria-model/src/test/java/fr/pizzeria/model/PizzaTest.java
@@ -1,5 +1,7 @@
 package fr.pizzeria.model;
 
+import java.math.BigDecimal;
+
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -7,15 +9,15 @@ public class PizzaTest {
 
 	@Test
 	public void testToString() {
-		Pizza p1 = new Pizza("IND", "L'indienne", 14.00, CategoriePizza.VIANDE);
-		Pizza p2 = new Pizza("ORI", "L'orientale", 13.50, CategoriePizza.VIANDE);
-		Pizza p3 = new Pizza("FRO", "La 4 fromages", 12.00, CategoriePizza.SANS_VIANDE);
-		Pizza p4 = new Pizza("REI", "La Reine", 11.50, CategoriePizza.VIANDE);
-		Pizza p5 = new Pizza("CAN", "La cannibale", 12.50, CategoriePizza.VIANDE);
-		Pizza p6 = new Pizza("SAU", "La saumonetta", 15.50, CategoriePizza.POISSON);
-		Pizza p7 = new Pizza("SAV", "La savoyarde", 13.00, CategoriePizza.VIANDE);
-		Pizza p8 = new Pizza("MAR", "Margherita", 14.00, CategoriePizza.SANS_VIANDE);
-		Pizza p9 = new Pizza("PEP", "Pépéroni", 12.50, CategoriePizza.VIANDE);
+		Pizza p1 = new Pizza("IND", "L'indienne", BigDecimal.valueOf(14.00), CategoriePizza.VIANDE);
+		Pizza p2 = new Pizza("ORI", "L'orientale", BigDecimal.valueOf(13.50), CategoriePizza.VIANDE);
+		Pizza p3 = new Pizza("FRO", "La 4 fromages", BigDecimal.valueOf(12.00), CategoriePizza.SANS_VIANDE);
+		Pizza p4 = new Pizza("REI", "La Reine", BigDecimal.valueOf(11.50), CategoriePizza.VIANDE);
+		Pizza p5 = new Pizza("CAN", "La cannibale", BigDecimal.valueOf(12.50), CategoriePizza.VIANDE);
+		Pizza p6 = new Pizza("SAU", "La saumonetta", BigDecimal.valueOf(15.50), CategoriePizza.POISSON);
+		Pizza p7 = new Pizza("SAV", "La savoyarde", BigDecimal.valueOf(13.00), CategoriePizza.VIANDE);
+		Pizza p8 = new Pizza("MAR", "Margherita", BigDecimal.valueOf(14.00), CategoriePizza.SANS_VIANDE);
+		Pizza p9 = new Pizza("PEP", "Pépéroni", BigDecimal.valueOf(12.50), CategoriePizza.VIANDE);
 
 		Assert.assertEquals("IND -> L'indienne [Viande] (14.0€)", p1.toString());
 		Assert.assertEquals("ORI -> L'orientale [Viande] (13.5€)", p2.toString());
